+++ components/SheetMusicApp/SheetMusicApp.js
@@ -6,6 +6,7 @@ import SheetMusic from '@slnsw/react-sheet-music';
 // import SheetMusic from '../SheetMusic';
 import Sidebar from '../Sidebar';
 import DXLabLogo from '../DXLabLogo';
+import MenuIconButton from '../MenuIconButton';
 
 import songs from './songs';
 import samples from './samples';
@@ -38,6 +39,19 @@ const SheetMusicApp = ({ slug, className }) => {
 
   // Check if all samples have been loaded
   const isSamplesLoaded = samplesStatus.every((status) => status === 'loaded');
+
+  // Override default body styles
+  React.useEffect(() => {
+    document
+      .getElementsByTagName('body')[0]
+      .classList.add(css.sheetMusicAppBody);
+
+    return () => {
+      document
+        .getElementsByTagName('body')[0]
+        .classList.remove(css.sheetMusicAppBody);
+    };
+  });
 
   const handleBeat = (beatNumber, totalBeats) => {
     if (beatNumber === totalBeats) {
@@ -86,6 +100,11 @@ const SheetMusicApp = ({ slug, className }) => {
     <div className={[css['sheet-music-app'], className || ''].join(' ')}>
       <header className={css.header}>
         <DXLabLogo />
+        <span className={css.headerDivider}></span>
+        <h1>
+          Virtu<strong>OSO</strong>
+        </h1>
+        <MenuIconButton className={css.menuIconButton} />
       </header>
 
       <div className={css.content}>