+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactor `Button` to `CTAButton`
 - Recreate `Button` component to be more generic
 - Add `MenuIconButton`
+- Add `eslint-plugin-css-modules`
 
 ## [1.8.1] - 2020-04-01